Adición de funcionalidad para el caso de uso 18

+++ SAIP/LogicaDeNegocio/Clases/Producto.cs
@@ -9,7 +9,7 @@ namespace LogicaDeNegocio.Clases
     public class Producto : Alimento
     {
         
-        public double CantidadEnInventario { get; set; }
+        public int CantidadEnInventario { get; set; }
         public string CodigoDeBarras { get; set; }
         public double Costo { get; set; }
         public string Creador { get; set; }
